adding trade policy button in trade advisor window
