+++ src/window/building/distribution.c
@@ -2,20 +2,17 @@
 
 #include "assets/assets.h"
 #include "building/building.h"
-#include "building/caravanserai.h"
 #include "building/dock.h"
 #include "building/market.h"
 #include "building/monument.h"
 #include "building/storage.h"
 #include "building/warehouse.h"
 #include "city/buildings.h"
-#include "city/finance.h"
 #include "city/military.h"
 #include "city/resource.h"
 #include "empire/city.h"
 #include "empire/object.h"
 #include "figure/figure.h"
-#include "game/resource.h"
 #include "graphics/generic_button.h"
 #include "graphics/image.h"
 #include "graphics/lang_text.h"
@@ -24,10 +21,9 @@
 #include "graphics/text.h"
 #include "graphics/window.h"
 #include "scenario/property.h"
-#include "sound/speech.h"
 #include "translation/translation.h"
-#include "window/action_popup.h"
 #include "window/building_info.h"
+#include "window/trade_policy.h"
 
 #include <math.h>
 
@@ -136,21 +132,6 @@ static struct {
     int dock_max_cities_visible;
     int caravanserai_focus_button_id;
 } data;
-
-static const option_menu_item caravanserai_policy_options[3] = {
-        {
-                0, TR_BUILDING_CARAVANSERAI_POLICY_1_TITLE, TR_BUILDING_CARAVANSERAI_POLICY_1,
-                0, "Areldir", "Econ_Logistics", "Trade Policy 1", 0
-        },
-        {
-                0, TR_BUILDING_CARAVANSERAI_POLICY_2_TITLE, TR_BUILDING_CARAVANSERAI_POLICY_2,
-                0, "Areldir", "Econ_Logistics", "Trade Policy 2", 0
-        },
-        {
-                0, TR_BUILDING_CARAVANSERAI_POLICY_3_TITLE, TR_BUILDING_CARAVANSERAI_POLICY_3,
-                0, "Areldir", "Econ_Logistics", "Trade Policy 3", 0
-        }
-};
 
 uint8_t warehouse_full_button_text[] = "32";
 uint8_t warehouse_3quarters_button_text[] = "24";
@@ -1175,57 +1156,14 @@ static void draw_policy_image_border(int x, int y, int focused)
     image_draw(id + 6 + focused, x, y + 5);
 }
 
-void window_building_draw_caravanserai_action(building_info_context *c)
-{
-
-    int policy = building_monument_module_type(BUILDING_CARAVANSERAI);
-    int policy_title = TR_BUILDING_CARAVANSERAI_NO_POLICY;
-
-    if (policy) {
-        policy_title = caravanserai_policy_options[policy - 1].header;
-    }
-
-    text_draw_multiline(translation_for(policy_title), c->x_offset + 160, c->y_offset + 160,
-        260, FONT_NORMAL_BLACK, 0);
-
-    if (policy) {
-        text_draw_multiline(translation_for(policy_title + 1), c->x_offset + 160, c->y_offset + 185,
-            260, FONT_NORMAL_BLACK, 0);
-
-        int policy_image = assets_get_image_id(assets_get_group_id((char *) caravanserai_policy_options[policy - 1].asset_author, (char *) caravanserai_policy_options[policy - 1].asset_name),
-            (char *) caravanserai_policy_options[policy - 1].asset_image_id);
-
-        image_draw(policy_image, c->x_offset + 32, c->y_offset + 150);
-    } else {
-        int policy_image = assets_get_image_id(assets_get_group_id("Areldir", "Econ_Logistics"), "Trade Policy");
-        image_draw(policy_image, c->x_offset + 32, c->y_offset + 150);
-    }
-
-}
-
 void window_building_draw_caravanserai_foreground(building_info_context *c)
 {
     draw_policy_image_border(c->x_offset + 32, c->y_offset + 150, data.caravanserai_focus_button_id == 1);
 }
 
-static void caravanserai_policy(int selection)
+void caravanserai_action(int param1, int param2)
 {
-    if (!selection) {
-        return;
-    }
-
-    sound_speech_play_file("wavs/market4.wav");
-    building *b = building_get(city_buildings_get_caravanserai());
-
-    city_finance_process_construction(TRADE_POLICY_COST);
-    building_monument_add_module(b, selection);
-}
-
-static void caravanserai_action(int param1, int param2)
-{
-    building *b = building_get(city_buildings_get_caravanserai());
-    int current_policy = b->data.monument.upgrades;
-    window_action_popup_show(TR_BUILDING_CARAVANSERAI_POLICY_TITLE, TR_BUILDING_CARAVANSERAI_POLICY_TEXT, caravanserai_policy_options, caravanserai_policy, current_policy);
+    window_policy_show();
 }
 
 void window_building_draw_caravanserai(building_info_context *c)
